+++ modulo3/labeddit/src/App.js
@@ -1,4 +1,4 @@
-import { Router } from './Router'
+import { Router } from './components/Router/Router'
  
 function App() {
   return (
