+++ modulo3/labeddit/src/components/CreateLogin/CreateLogin.jsx
@@ -1,37 +1,16 @@
-import styled from 'styled-components'
 import { useNavigate } from 'react-router-dom'
 import { useState } from 'react'
 import axios from "axios";
-import { AppHeader } from './styled';
+import { AppHeader, Container, ContainerButton } from '../styled';
+import { ToSee } from '../Hooks/switcher'
+import { goToLogin } from '../Router/links';
 
-const ContainerButton = styled.div`
-button {
-    border-radius: 10px;
-    margin: 20px 20px;
-    padding: 10px;
-    border: none;
-    cursor: pointer;
-    &:hover{
-        background-color:gray;
-    }  
-}
-`
-const Container = styled.div`
-    margin: 25px;
-    input {
-        border-radius: 10px;
-        border: none;
-        padding: 5px;
-    }
-    div {
-        cursor: pointer
-    }
-`
 export default function CreateLogin() {
+    const navigate = useNavigate()
     const [nick, setNick] = useState('')
     const [email, setEmail] = useState('')
     const [passw, setPassw] = useState('')
-    const [seePass, setSeePass] = useState('password')
+    const [ seePass, switcher ] = ToSee()
     const [pag, setPag] = useState(false)
     const onChangeNick = (e) => {
         setNick(e.target.value)
@@ -49,21 +28,7 @@ export default function CreateLogin() {
             })
             .catch((err) => alert(err.statusText, 'Senha incorreta'))
     }
-    const Create = useNavigate()
-    const goToCreateAccount = () => {
-        Create('/')
-    }
-    const Voltar = useNavigate()
-    const goToLogin = () => {
-        Voltar('/')
-    }
-    const switcher = () => {
-        if (seePass == 'password') {
-            setSeePass('text')
-        } else {
-            setSeePass('password')
-        }
-    }
+
     return (
         <AppHeader>
             <div>
@@ -73,7 +38,7 @@ export default function CreateLogin() {
                     <Container><input onChange={onChangeMail} type={'Text'} placeholder='Email do usuario' ></input></Container>
                     <Container><input onChange={onChangePassw} type={seePass} placeholder='Digite sua senha' ></input>
                     <div onClick={switcher}>ver senha</div></Container>
-                    <button onClick={goToLogin}>Voltar</button>
+                    <button onClick={() => goToLogin(navigate)}>Voltar</button>
                     <button onClick={() => alert('Conta criada')}>Criar</button>
                 </ContainerButton>
             </div>
