--- modulo3/labeddit/src/components/Feed.jsx
@@ -1,10 +1,0 @@
-
-
-export default function Feed() {
-
-    return (
-        
-        <div>----Principais noticias e materias----<br/>-----Pagina em construçao-----</div>
-    )
-
-}
