+++ modulo3/labeddit/src/components/Feed/Feed.jsx
@@ -1,0 +1,10 @@
+
+
+export default function Feed() {
+
+    return (
+        
+        <div>----TODOS OS POSTS----<br/>-----Pagina em construçao-----</div>
+    )
+
+}
