+++ modulo3/labeddit/src/components/Router/links.jsx
@@ -1,0 +1,13 @@
+
+export const goToLogin = (x) => {
+   return x('/')
+}
+
+export const goToCreateAccount = (x) => {
+   return x('/CreateAccount')
+}
+
+export const goToFeed = (x) => {
+    x('/Feed')
+}
+
